requirements.txt
google-cloud-bigquery
google-cloud-pubsub
google-cloud-storage
python-dotenv